move Python to own dir
