+++ packages/quicktype-core/src/language/Python/JSONPythonRenderer.ts
@@ -1,31 +1,9 @@
-import {
-    arrayIntercalate,
-    iterableFirst,
-    iterableSome,
-    mapSortBy,
-    mapUpdateInto,
-    setUnionInto
-} from "collection-utils";
-import unicode from "unicode-properties";
+import { arrayIntercalate } from "collection-utils";
 
-import { ConvenienceRenderer, type ForbiddenWordsInfo, topLevelNameOrder } from "../ConvenienceRenderer";
-import { DependencyName, type Name, type Namer, funPrefixNamer } from "../Naming";
-import { type RenderContext } from "../Renderer";
-import { BooleanOption, EnumOption, type Option, type OptionValues, getOptionValues } from "../RendererOptions";
-import { type MultiWord, type Sourcelike, modifySource, multiWord, parenIfNeeded, singleWord } from "../Source";
-import {
-    type WordStyle,
-    allLowerWordStyle,
-    allUpperWordStyle,
-    combineWords,
-    firstUpperWordStyle,
-    originalWord,
-    splitIntoWords,
-    stringEscape,
-    utf16LegalizeCharacters
-} from "../support/Strings";
-import { assertNever, defined, panic } from "../support/Support";
-import { TargetLanguage } from "../TargetLanguage";
+import { topLevelNameOrder } from "../../ConvenienceRenderer";
+import { DependencyName, type Name, funPrefixNamer } from "../../Naming";
+import { type MultiWord, type Sourcelike, multiWord, parenIfNeeded, singleWord } from "../../Source";
+import { assertNever, defined, panic } from "../../support/Support";
 import {
     ChoiceTransformer,
     DecodingChoiceTransformer,
@@ -36,512 +14,13 @@ import {
     type Transformer,
     UnionInstantiationTransformer,
     UnionMemberMatchTransformer,
-    followTargetType,
     transformationForType
-} from "../Transformers";
-import {
-    type ClassProperty,
-    ClassType,
-    EnumType,
-    type PrimitiveStringTypeKind,
-    type TransformedStringTypeKind,
-    type Type,
-    UnionType
-} from "../Type";
-import { type StringTypeMapping } from "../TypeBuilder";
-import { type FixMeOptionsAnyType, type FixMeOptionsType } from "../types";
-import { matchType, nullableFromUnion, removeNullFromUnion } from "../TypeUtils";
-
-const forbiddenTypeNames = [
-    "Any",
-    "True",
-    "False",
-    "None",
-    "Enum",
-    "List",
-    "Dict",
-    "Optional",
-    "Union",
-    "Iterable",
-    "Type",
-    "TypeVar",
-    "T",
-    "EnumT"
-];
-const forbiddenPropertyNames = [
-    "and",
-    "as",
-    "assert",
-    "async",
-    "await",
-    "bool",
-    "break",
-    "class",
-    "continue",
-    "datetime",
-    "def",
-    "del",
-    "dict",
-    "elif",
-    "else",
-    "except",
-    "finally",
-    "float",
-    "for",
-    "from",
-    "global",
-    "if",
-    "import",
-    "in",
-    "int",
-    "is",
-    "lambda",
-    "nonlocal",
-    "not",
-    "or",
-    "pass",
-    "print",
-    "raise",
-    "return",
-    "self",
-    "str",
-    "try",
-    "while",
-    "with",
-    "yield"
-];
-
-export interface PythonFeatures {
-    dataClasses: boolean;
-    typeHints: boolean;
-}
-
-export const pythonOptions = {
-    features: new EnumOption<PythonFeatures>(
-        "python-version",
-        "Python version",
-        [
-            ["3.5", { typeHints: false, dataClasses: false }],
-            ["3.6", { typeHints: true, dataClasses: false }],
-            ["3.7", { typeHints: true, dataClasses: true }]
-        ],
-        "3.6"
-    ),
-    justTypes: new BooleanOption("just-types", "Classes only", false),
-    nicePropertyNames: new BooleanOption("nice-property-names", "Transform property names to be Pythonic", true)
-};
-
-export class PythonTargetLanguage extends TargetLanguage {
-    protected getOptions(): Array<Option<FixMeOptionsAnyType>> {
-        return [pythonOptions.features, pythonOptions.justTypes, pythonOptions.nicePropertyNames];
-    }
-
-    public get stringTypeMapping(): StringTypeMapping {
-        const mapping: Map<TransformedStringTypeKind, PrimitiveStringTypeKind> = new Map();
-        const dateTimeType = "date-time";
-        mapping.set("date", dateTimeType);
-        mapping.set("time", dateTimeType);
-        mapping.set("date-time", dateTimeType);
-        mapping.set("uuid", "uuid");
-        mapping.set("integer-string", "integer-string");
-        mapping.set("bool-string", "bool-string");
-        return mapping;
-    }
-
-    public get supportsUnionsWithBothNumberTypes(): boolean {
-        return true;
-    }
-
-    public get supportsOptionalClassProperties(): boolean {
-        return false;
-    }
-
-    public needsTransformerForType(t: Type): boolean {
-        if (t instanceof UnionType) {
-            return iterableSome(t.members, m => this.needsTransformerForType(m));
-        }
-
-        return t.kind === "integer-string" || t.kind === "bool-string";
-    }
-
-    protected makeRenderer(renderContext: RenderContext, untypedOptionValues: FixMeOptionsType): PythonRenderer {
-        const options = getOptionValues(pythonOptions, untypedOptionValues);
-        if (options.justTypes) {
-            return new PythonRenderer(this, renderContext, options);
-        } else {
-            return new JSONPythonRenderer(this, renderContext, options);
-        }
-    }
-}
-
-function isNormalizedStartCharacter3(utf16Unit: number): boolean {
-    // FIXME: add Other_ID_Start - https://docs.python.org/3/reference/lexical_analysis.html#identifiers
-    const category: string = unicode.getCategory(utf16Unit);
-    return ["Lu", "Ll", "Lt", "Lm", "Lo", "Nl"].includes(category);
-}
-
-function isNormalizedPartCharacter3(utf16Unit: number): boolean {
-    // FIXME: add Other_ID_Continue - https://docs.python.org/3/reference/lexical_analysis.html#identifiers
-    if (isNormalizedStartCharacter3(utf16Unit)) return true;
-    const category: string = unicode.getCategory(utf16Unit);
-    return ["Mn", "Mc", "Nd", "Pc"].includes(category);
-}
-
-function isStartCharacter3(utf16Unit: number): boolean {
-    const s = String.fromCharCode(utf16Unit).normalize("NFKC");
-    const l = s.length;
-    if (l === 0 || !isNormalizedStartCharacter3(s.charCodeAt(0))) return false;
-    for (let i = 1; i < l; i++) {
-        if (!isNormalizedPartCharacter3(s.charCodeAt(i))) return false;
-    }
-
-    return true;
-}
-
-function isPartCharacter3(utf16Unit: number): boolean {
-    const s = String.fromCharCode(utf16Unit).normalize("NFKC");
-    const l = s.length;
-    for (let i = 0; i < l; i++) {
-        if (!isNormalizedPartCharacter3(s.charCodeAt(i))) return false;
-    }
-
-    return true;
-}
-
-const legalizeName3 = utf16LegalizeCharacters(isPartCharacter3);
-
-function classNameStyle(original: string): string {
-    const words = splitIntoWords(original);
-    return combineWords(
-        words,
-        legalizeName3,
-        firstUpperWordStyle,
-        firstUpperWordStyle,
-        allUpperWordStyle,
-        allUpperWordStyle,
-        "",
-        isStartCharacter3
-    );
-}
-
-function getWordStyle(uppercase: boolean, forceSnakeNameStyle: boolean): WordStyle {
-    if (!forceSnakeNameStyle) {
-        return originalWord;
-    }
-
-    return uppercase ? allUpperWordStyle : allLowerWordStyle;
-}
-
-function snakeNameStyle(original: string, uppercase: boolean, forceSnakeNameStyle: boolean): string {
-    const wordStyle = getWordStyle(uppercase, forceSnakeNameStyle);
-    const separator = forceSnakeNameStyle ? "_" : "";
-    const words = splitIntoWords(original);
-    return combineWords(words, legalizeName3, wordStyle, wordStyle, wordStyle, wordStyle, separator, isStartCharacter3);
-}
-
-export class PythonRenderer extends ConvenienceRenderer {
-    private readonly imports: Map<string, Set<string>> = new Map();
-
-    private readonly declaredTypes: Set<Type> = new Set();
-
-    public constructor(
-        targetLanguage: TargetLanguage,
-        renderContext: RenderContext,
-        protected readonly pyOptions: OptionValues<typeof pythonOptions>
-    ) {
-        super(targetLanguage, renderContext);
-    }
-
-    protected forbiddenNamesForGlobalNamespace(): string[] {
-        return forbiddenTypeNames;
-    }
-
-    protected forbiddenForObjectProperties(_: ClassType, _classNamed: Name): ForbiddenWordsInfo {
-        return { names: forbiddenPropertyNames, includeGlobalForbidden: false };
-    }
-
-    protected makeNamedTypeNamer(): Namer {
-        return funPrefixNamer("type", classNameStyle);
-    }
-
-    protected namerForObjectProperty(): Namer {
-        return funPrefixNamer("property", s => snakeNameStyle(s, false, this.pyOptions.nicePropertyNames));
-    }
-
-    protected makeUnionMemberNamer(): null {
-        return null;
-    }
-
-    protected makeEnumCaseNamer(): Namer {
-        return funPrefixNamer("enum-case", s => snakeNameStyle(s, true, this.pyOptions.nicePropertyNames));
-    }
-
-    protected get commentLineStart(): string {
-        return "# ";
-    }
-
-    protected emitDescriptionBlock(lines: Sourcelike[]): void {
-        if (lines.length === 1) {
-            const docstring = modifySource(content => {
-                if (content.endsWith('"')) {
-                    return content.slice(0, -1) + '\\"';
-                }
-
-                return content;
-            }, lines[0]);
-            this.emitComments([{ customLines: [docstring], lineStart: '"""', lineEnd: '"""' }]);
-        } else {
-            this.emitCommentLines(lines, {
-                firstLineStart: '"""',
-                lineStart: "",
-                afterComment: '"""'
-            });
-        }
-    }
-
-    protected get needsTypeDeclarationBeforeUse(): boolean {
-        return true;
-    }
-
-    protected canBeForwardDeclared(t: Type): boolean {
-        const kind = t.kind;
-        return kind === "class" || kind === "enum";
-    }
-
-    protected emitBlock(line: Sourcelike, f: () => void): void {
-        this.emitLine(line);
-        this.indent(f);
-    }
-
-    protected string(s: string): Sourcelike {
-        const openQuote = '"';
-        return [openQuote, stringEscape(s), '"'];
-    }
-
-    protected withImport(module: string, name: string): Sourcelike {
-        if (this.pyOptions.features.typeHints || module !== "typing") {
-            // FIXME: This is ugly.  We should rather not generate that import in the first
-            // place, but right now we just make the type source and then throw it away.  It's
-            // not a performance issue, so it's fine, I just bemoan this special case, and
-            // potential others down the road.
-            mapUpdateInto(this.imports, module, s => (s ? setUnionInto(s, [name]) : new Set([name])));
-        }
-
-        return name;
-    }
-
-    protected withTyping(name: string): Sourcelike {
-        return this.withImport("typing", name);
-    }
-
-    protected namedType(t: Type): Sourcelike {
-        const name = this.nameForNamedType(t);
-        if (this.declaredTypes.has(t)) return name;
-        return ["'", name, "'"];
-    }
-
-    protected pythonType(t: Type, _isRootTypeDef = false): Sourcelike {
-        const actualType = followTargetType(t);
-
-        return matchType<Sourcelike>(
-            actualType,
-            _anyType => this.withTyping("Any"),
-            _nullType => "None",
-            _boolType => "bool",
-            _integerType => "int",
-            _doubletype => "float",
-            _stringType => "str",
-            arrayType => [this.withTyping("List"), "[", this.pythonType(arrayType.items), "]"],
-            classType => this.namedType(classType),
-            mapType => [this.withTyping("Dict"), "[str, ", this.pythonType(mapType.values), "]"],
-            enumType => this.namedType(enumType),
-            unionType => {
-                const [hasNull, nonNulls] = removeNullFromUnion(unionType);
-                const memberTypes = Array.from(nonNulls).map(m => this.pythonType(m));
-
-                if (hasNull !== null) {
-                    let rest: string[] = [];
-                    if (!this.getAlphabetizeProperties() && this.pyOptions.features.dataClasses && _isRootTypeDef) {
-                        // Only push "= None" if this is a root level type def
-                        //   otherwise we may get type defs like List[Optional[int] = None]
-                        //   which are invalid
-                        rest.push(" = None");
-                    }
-
-                    if (nonNulls.size > 1) {
-                        this.withImport("typing", "Union");
-                        return [
-                            this.withTyping("Optional"),
-                            "[Union[",
-                            arrayIntercalate(", ", memberTypes),
-                            "]]",
-                            ...rest
-                        ];
-                    } else {
-                        return [this.withTyping("Optional"), "[", defined(iterableFirst(memberTypes)), "]", ...rest];
-                    }
-                } else {
-                    return [this.withTyping("Union"), "[", arrayIntercalate(", ", memberTypes), "]"];
-                }
-            },
-            transformedStringType => {
-                if (transformedStringType.kind === "date-time") {
-                    return this.withImport("datetime", "datetime");
-                }
-
-                if (transformedStringType.kind === "uuid") {
-                    return this.withImport("uuid", "UUID");
-                }
-
-                return panic(`Transformed type ${transformedStringType.kind} not supported`);
-            }
-        );
-    }
-
-    protected declarationLine(t: Type): Sourcelike {
-        if (t instanceof ClassType) {
-            return ["class ", this.nameForNamedType(t), ":"];
-        }
-
-        if (t instanceof EnumType) {
-            return ["class ", this.nameForNamedType(t), "(", this.withImport("enum", "Enum"), "):"];
-        }
-
-        return panic(`Can't declare type ${t.kind}`);
-    }
-
-    protected declareType<T extends Type>(t: T, emitter: () => void): void {
-        this.emitBlock(this.declarationLine(t), () => {
-            this.emitDescription(this.descriptionForType(t));
-            emitter();
-        });
-        this.declaredTypes.add(t);
-    }
-
-    protected emitClassMembers(t: ClassType): void {
-        if (this.pyOptions.features.dataClasses) return;
-
-        const args: Sourcelike[] = [];
-        this.forEachClassProperty(t, "none", (name, _, cp) => {
-            args.push([name, this.typeHint(": ", this.pythonType(cp.type))]);
-        });
-        this.emitBlock(
-            ["def __init__(self, ", arrayIntercalate(", ", args), ")", this.typeHint(" -> None"), ":"],
-            () => {
-                if (args.length === 0) {
-                    this.emitLine("pass");
-                } else {
-                    this.forEachClassProperty(t, "none", name => {
-                        this.emitLine("self.", name, " = ", name);
-                    });
-                }
-            }
-        );
-    }
-
-    protected typeHint(...sl: Sourcelike[]): Sourcelike {
-        if (this.pyOptions.features.typeHints) {
-            return sl;
-        }
-
-        return [];
-    }
-
-    protected typingDecl(name: Sourcelike, type: string): Sourcelike {
-        return [name, this.typeHint(": ", this.withTyping(type))];
-    }
-
-    protected typingReturn(type: string): Sourcelike {
-        return this.typeHint(" -> ", this.withTyping(type));
-    }
-
-    protected sortClassProperties(
-        properties: ReadonlyMap<string, ClassProperty>,
-        propertyNames: ReadonlyMap<string, Name>
-    ): ReadonlyMap<string, ClassProperty> {
-        if (this.pyOptions.features.dataClasses) {
-            return mapSortBy(properties, (p: ClassProperty) => {
-                return (p.type instanceof UnionType && nullableFromUnion(p.type) != null) || p.isOptional ? 1 : 0;
-            });
-        } else {
-            return super.sortClassProperties(properties, propertyNames);
-        }
-    }
-
-    protected emitClass(t: ClassType): void {
-        if (this.pyOptions.features.dataClasses) {
-            this.emitLine("@", this.withImport("dataclasses", "dataclass"));
-        }
-
-        this.declareType(t, () => {
-            if (this.pyOptions.features.typeHints) {
-                if (t.getProperties().size === 0) {
-                    this.emitLine("pass");
-                } else {
-                    this.forEachClassProperty(t, "none", (name, jsonName, cp) => {
-                        this.emitLine(name, this.typeHint(": ", this.pythonType(cp.type, true)));
-                        this.emitDescription(this.descriptionForClassProperty(t, jsonName));
-                    });
-                }
-
-                this.ensureBlankLine();
-            }
-
-            this.emitClassMembers(t);
-        });
-    }
-
-    protected emitEnum(t: EnumType): void {
-        this.declareType(t, () => {
-            this.forEachEnumCase(t, "none", (name, jsonName) => {
-                this.emitLine([name, " = ", this.string(jsonName)]);
-            });
-        });
-    }
-
-    protected emitImports(): void {
-        this.imports.forEach((names, module) => {
-            this.emitLine("from ", module, " import ", Array.from(names).join(", "));
-        });
-    }
-
-    protected emitSupportCode(): void {
-        return;
-    }
-
-    protected emitClosingCode(): void {
-        return;
-    }
-
-    protected emitSourceStructure(_givenOutputFilename: string): void {
-        const declarationLines = this.gatherSource(() => {
-            this.forEachNamedType(
-                ["interposing", 2],
-                (c: ClassType) => this.emitClass(c),
-                e => this.emitEnum(e),
-                _u => {
-                    return;
-                }
-            );
-        });
-
-        const closingLines = this.gatherSource(() => this.emitClosingCode());
-        const supportLines = this.gatherSource(() => this.emitSupportCode());
-
-        if (this.leadingComments !== undefined) {
-            this.emitComments(this.leadingComments);
-        }
-
-        this.ensureBlankLine();
-        this.emitImports();
-        this.ensureBlankLine(2);
-        this.emitGatheredSource(supportLines);
-        this.ensureBlankLine(2);
-        this.emitGatheredSource(declarationLines);
-        this.ensureBlankLine(2);
-        this.emitGatheredSource(closingLines);
-    }
-}
+} from "../../Transformers";
+import { type ClassType, type Type } from "../../Type";
+import { matchType } from "../../TypeUtils";
+
+import { PythonRenderer } from "./PythonRenderer";
+import { snakeNameStyle } from "./utils";
 
 export type ConverterFunction =
     | "none"
