+++ packages/quicktype-core/src/language/Python/constants.ts
@@ -1,0 +1,59 @@
+export const forbiddenTypeNames = [
+    "Any",
+    "True",
+    "False",
+    "None",
+    "Enum",
+    "List",
+    "Dict",
+    "Optional",
+    "Union",
+    "Iterable",
+    "Type",
+    "TypeVar",
+    "T",
+    "EnumT"
+] as const;
+
+export const forbiddenPropertyNames = [
+    "and",
+    "as",
+    "assert",
+    "async",
+    "await",
+    "bool",
+    "break",
+    "class",
+    "continue",
+    "datetime",
+    "def",
+    "del",
+    "dict",
+    "elif",
+    "else",
+    "except",
+    "finally",
+    "float",
+    "for",
+    "from",
+    "global",
+    "if",
+    "import",
+    "in",
+    "int",
+    "is",
+    "lambda",
+    "nonlocal",
+    "not",
+    "or",
+    "pass",
+    "print",
+    "raise",
+    "return",
+    "self",
+    "str",
+    "try",
+    "while",
+    "with",
+    "yield"
+] as const;
