+++ packages/quicktype-core/src/language/Python/index.ts
@@ -1,0 +1,3 @@
+export { PythonTargetLanguage, pythonOptions } from "./language";
+export { PythonRenderer } from "./PythonRenderer";
+export { JSONPythonRenderer } from "./JSONPythonRenderer";
